Move all printing methods to single file
